pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dinov3easy"
version = "2025.09.03-001"
description = "A DINOv3 wrapper that simplifies usage and visualization."
readme = "README.md"
authors = [
    { name = "Matteo Leccardi", email = "matteo.leccardi@polimi.it" }
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pillow",
    "torch",
    "torchmetrics",
    "torchvision",
    "scikit-learn",
    "matplotlib",
    "ftfy",
    "omegaconf",
    "regex",
    "submitit",
    "termcolor",
]
